:art: Reorganize internals WIP
