+++ lib/react/cells/tags-component.js
@@ -39,7 +39,7 @@ export default React.createClass({
     })
   },
 
-  componentDidUnmount () {
+  componentWillUnmount () {
     this._unsub()
   },
