+++ lib/ui-atom-adapter.js
@@ -1,22 +1,25 @@
 'use babel'
 
-import {React} from 'react-for-atom'
+import {React, ReactDOM} from 'react-for-atom'
+import R from 'ramda'
 import Bacon from 'baconjs'
 import * as atoms from './atom-streams'
 import DisposableValues from './disposable-values'
-import DateTimeComponent from './react/cells/date-time-component'
-import SummaryComponent from './react/cells/summary-component'
-import TagsComponent from './react/cells/tags-component'
-import UI from './ui'
+import PanelComponent from './react/panel-component'
+import newSelectedProps from './new-selected-props'
+import newColumns from './new-columns'
 
 // Adapter for Atom-land
 // Basically all atom.* usages that contains side-effects related to the UI should go here.
 export default class UIAtomAdapter {
 
   constructor (project) {
-    const panelDomNode = document.createElement('div')
-    this._atomPanel = atom.workspace.addTopPanel({item: panelDomNode})
+    this._domNode = document.createElement('div')
+    this._atomPanel = atom.workspace.addTopPanel({item: this._domNode})
     this._previewEditor = null
+
+    const keyDownBus = new Bacon.Bus()
+    const selectIndexBus = new Bacon.Bus()
 
     const closePreviewEditorStream = atoms
       .createStream(atom.workspace, 'onDidDestroyPaneItem')
@@ -28,69 +31,45 @@ export default class UIAtomAdapter {
     const focusOnSearchStream = Bacon
       .mergeAll(
         focusCmdStream,
+        selectIndexBus,
         togglePanelCmdStream,
         toggleAtomWinCmdStream,
+        project.isLoadingFilesProp.changes(),
         // Focus on search when panel changes visibility (to visible)
         atoms.createStream(this._atomPanel, 'onDidChangeVisible').filter(isVisible => isVisible),
         // Focus on search when active editor gets focus (e.g. on click or other model cancels)
         atoms.createStream(atom.workspace, 'onDidChangeActivePaneItem').filter(item => item === this._previewEditor)
       )
       .debounce(50)
-      .merge(project.isLoadingFilesProp.changes()) // placed after debounce, to trigger immeditately
 
-    const columns = []
-    columns.push({
-      title: 'Name',
-      width: 50,
-      createCell: (file, state) => {
-        return <SummaryComponent key='name' file={file}
-        searchStr={state.searchStr} tokens={state.results.tokens} />
-      }
-    })
-    if (project.darwin) {
-      const editTagsStream = atoms.createCommandStream('atom-workspace', 'textual-velocity:edit-tags')
-      columns.push({
-        title: 'Tags',
-        width: 20,
-        createCell: (file, state, isSelected) => {
-          const saveTags = tags => project.darwin.setTags(file.path, tags.trim().split(' '))
-          return (
-            <TagsComponent key='tags' tags={file.tags} isSelected={isSelected} saveTags={saveTags}
-              editTagsStream={editTagsStream} />
-          )
-        }
-      })
-    }
-    columns.push({
-      title: 'Date modified',
-      width: 15,
-      createCell: file => {
-        return <DateTimeComponent key='mtime' time={file.stat.mtime} />
-      }
-    })
-    columns.push({
-      title: 'Date created',
-      width: 15,
-      createCell: file => {
-        return <DateTimeComponent key='birhtime' time={file.stat.birthtime} />
-      }
-    })
-
-    const ui = new UI({
-      domNode: panelDomNode,
-      isLoadingFilesProp: project.isLoadingFilesProp,
+    const {selectedIndexStream, selectedFileProp} = newSelectedProps({
       filesProp: project.filesProp,
       resultsProp: project.resultsProp,
-      columns: columns,
-      panelHeightStream: atoms.createConfigStream('textual-velocity.panelHeight'),
-      focusOnSearchStream: focusOnSearchStream,
-      deselectStream: Bacon.mergeAll(
-        closePreviewEditorStream,
-        project.searchBus,
-      )})
+      keyDownBus: keyDownBus,
+      selectIndexBus: selectIndexBus,
+      deselectStream: Bacon
+        .mergeAll(
+          closePreviewEditorStream,
+          project.searchBus,
+        )
+    })
+
+    const reactPanel = ReactDOM.render(
+      <PanelComponent
+        filesProp={project.filesProp}
+        resultsProp={project.resultsProp}
+        isLoadingFilesProp={project.isLoadingFilesProp}
+        selectedIndexStream={selectedIndexStream}
+        focusOnSearchStream={focusOnSearchStream}
+        keyDownBus={keyDownBus}
+        selectIndexBus={selectIndexBus}
+        panelHeightStream={atoms.createConfigStream('textual-velocity.panelHeight')}
+        columns={newColumns(project.darwin)}
+      />, this._domNode)
 
     // Make sure search changes are wired back to project's search bus
-    project.searchBus.plug(ui.searchProp.toEventStream())
+    // TODO can remove?
+    project.searchBus.plug(reactPanel.searchProp.toEventStream())
 
     // Focus on panel when command is triggered
     this._disposables = new DisposableValues()
@@ -112,7 +91,7 @@ export default class UIAtomAdapter {
 
     // Persist the panel height when changed
     this._disposables.add(
-      ui.panelHeightProp
+      reactPanel.panelHeightProp
         .debounce(500)
         .onValue(h => {
           atom.config.set('textual-velocity.panelHeight', h)
@@ -120,9 +99,10 @@ export default class UIAtomAdapter {
 
     // Preview or close file when selection changes
     this._disposables.add(
-      ui.selectedFileProp
+      selectedFileProp
         // Open a file is not instant, so debounce to avoid sluggishness on rapid changes (e.g. pressing down/up)
         .debounce(50)
+        // TODO extract and test separately? "this._"-usage is a little problematic
         .onValue(selectedFile => {
           if (selectedFile) {
             if (this._previewEditor) {
@@ -156,17 +136,35 @@ export default class UIAtomAdapter {
         this._previewEditor = null
       }))
 
+    // TODO how/what should these concepts be named/done?
+    const enterKeyStream = keyDownBus
+      .filter(R.propEq('keyCode', 13))
+      .doLog('[UI_enterKeyStream]')
+    const selectedFilePathOnEnterStream = selectedFileProp
+      .map('.path')
+      .skipDuplicates()
+      .sampledBy(enterKeyStream)
+      .doLog('[UI_selectedFilePathOnEnterStream]')
+
     // Open selected file path when item is selected
     this._disposables.add(
-      ui.openFileStream
+      selectedFilePathOnEnterStream
+        .filter(R.identity)
         .onValue(path => {
           atom.workspace.open(path)
         }))
 
+    // TODO newFilePathProp should come from here?
+    const createFileStream = Bacon
+      .when(
+        [selectedFilePathOnEnterStream.filter(R.not), reactPanel.searchProp],
+        (_, searchStr) => {
+          return searchStr.trim()
+        }
+      )
     this._disposables.add(
-      project
-        .newFilePathProp
-        .sampledBy(ui.createFileStream)
+      project.newFilePathProp
+        .sampledBy(createFileStream)
         .onValue(path => {
           atom.workspace.open(path)
         }))
@@ -186,13 +184,15 @@ export default class UIAtomAdapter {
             atom.focus()
             this._atomPanel.show()
           }
-        })
-    )
+        }))
   }
 
   dispose () {
     this._atomPanel.destroy()
     this._atomPanel = null
+
+    ReactDOM.unmountComponentAtNode(this._domNode)
+    this._domNode = null
 
     this._disposables.dispose()
     this._disposables = null
